Decoding user codes 35-39, 43+ for panelStatus 0x05, 0x17, 0x18
Little messy but it works for *1, *2, *3, *5, *6, *7/Access codes codes 43-75. Arming/disarming is decoded but not implemented yet.
Decoding for user codes 75 - 95 is included but not tested (I dont have PC1864) but range matches fine so it should be correct.
It was mainly for testing and to learn/gain some experience with dsckeybusinterface structure. Code should be optimized.
Also, I had trouble getting user codes 35-39 to work because printPanelAccessCode increases user code by 5, needed for cmds on older generation.